+++ CHANGELOG.md
@@ -21,7 +21,7 @@ Types of changes
 ## [Unreleased] - ...
 -->
 
-## [Unreleased] - ...
+## [2.10.1] - 2021-03-14
 
 ## Added